src/gpu_vllm/pyproject.toml
[project]
name = "llmrec-gpu"
version = "2.0.0"
description = "GPU pipeline for history-shock 2x2 experiment"
requires-python = ">=3.10"
dependencies = [
    "vllm>=0.4",
    "transformers>=4.40",
    "numpy",
    "pandas",
    "scipy",
    "openai>=1.0",
    "tenacity",
    "python-dotenv",
    "tqdm",
]
